+++ data/meson.build
@@ -68,3 +68,11 @@ if compile_schemas.found()
 endif
 
 subdir('icons')
+
+gnome = import('gnome')
+
+gnome.post_install(
+  glib_compile_schemas: true,
+  gtk_update_icon_cache: true,
+  update_desktop_database: true,
+)
