+++ meson.build
@@ -8,14 +8,29 @@ project('logistics',
 
 i18n = import('i18n')
 
-gnome = import('gnome')
 
+pkgdatadir = join_paths(get_option('prefix'), get_option('datadir'), meson.project_name())
 subdir('data')
-subdir('src')
-subdir('po')
 
-gnome.post_install(
-  glib_compile_schemas: true,
-  gtk_update_icon_cache: true,
-  update_desktop_database: true,
+
+python = import('python')
+
+conf = configuration_data()
+conf.set('PYTHON', python.find_installation('python3').path())
+conf.set('VERSION', meson.project_version())
+conf.set('localedir', join_paths(get_option('prefix'), get_option('localedir')))
+conf.set('pkgdatadir', pkgdatadir)
+
+
+configure_file(
+  input: 'logistics.in',
+  output: 'logistics',
+  configuration: conf,
+  install: true,
+  install_dir: get_option('bindir')
 )
+
+
+install_subdir('logistics', install_dir: pkgdatadir)
+
+subdir('po')
